+++ cdlclient/__init__.py
@@ -16,7 +16,7 @@ application through XML-RPC protocol.
 from cdlclient.baseproxy import SimpleBaseProxy
 from cdlclient.remote import SimpleRemoteProxy
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 __docurl__ = "https://cdlclient.readthedocs.io/en/latest/"
 __homeurl__ = "https://github.com/Codra-Ingenierie-Informatique/DataLabSimpleClient/"
 __supporturl__ = "https://github.com/Codra-Ingenierie-Informatique/DataLabSimpleClient/issues/new/choose"