+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Changes:
   * `get_object_shapes` method now takes either object number, UUID or a title
   * Removed deprecated `get_object_from_uuid` and `get_object_from_title` methods
 
+* Simplified DataLab object model:
+  * Added `SignalObj.uuid` item
+  * Added `ImageObj.uuid` item
+
 ## Version 0.2.0 ##
 
 Changes: